refactor(vrt): separate editor-validity into its own directory
Move editor-validity.test.ts out of vrt/libreoffice/ and split its
fixture generation (editor-validity pairs + basic-shapes source) into
vrt/editor-validity/create_fixtures.py, so renderer regression VRT and
editor validity checks no longer share files or fixture directories.
